Correlation Analysis
The trend women postponing the time getting pregnant happened in many countries these years. Detecting the correlation between women’s maternal age and common influencing factors is meaningful for both policy makers to formulate incentive policies and the public to have the awareness of planning for the time of childbearing. The correlation analysis used data mining with Python and Python packages. International data over the time period from 1998 to 2016 are selected.